Massive code cleanup
I used IntelliJ IDEA's code analysis feature to help find warnings and bugs, and manually went through fixing them. The only warnings I let IDEA fix automatically were unused imports and most redundant casts.
ItemStacks have been converted to use ItemStack.EMPTY instead of null, to conform to new standards.
Annotations have been added to a LOT of the code to help devs determine when null can and can't be used, however they're nowhere near all-encompassing.
The inheritance of the BlockAdvancedBipropellantRocketMotor class was rearranged to prevent a potential mis-classification of them as monopropellant motors.
Vague fields and variables have been renamed where I found them, such as i -> itemStack or obj -> spaceObject. This excludes math variables that seemed intentionally named, and standardized names like loop control variables.
Lang files have been cleaned up to the best of my ability, but the non-en_US languages still have a lot of missing translation keys. Probably best for someone who speaks those languages to add the translations, rather than me with Google Translate.
There were some recipes that had invalid json, so I fixed them. This resulted in a few of the recipes changing slightly.
This commit should fix a lot of NPEs, though I didn't get them all.

I highly recommend one of the original devs merge this and also run the source through their own IDE's code analysis, since there are still a lot of warnings I couldn't or didn't fix, and a large amount of them could be or turn into bugs.
The ItemStack changes affected some API methods. Mods developed against Advanced Rocketry should use builds containing this commit in the future, to prevent them from sending null ItemStacks. I did still split any changes to method names and prototypes into a separate branch that can be merged or ignored.
This commit requires a sister-commit for libVulpes, since the ItemStack refactoring affected methods there too.

+++ src/main/java/zmaster587/advancedRocketry/block/BlockOrbitalLaserDrill.java
@@ -16,7 +16,7 @@ import java.util.Random;
 public class BlockOrbitalLaserDrill extends BlockMultiblockMachine {
 
 	public BlockOrbitalLaserDrill() {
-		super(TileOrbitalLaserDrill.class, (int)GuiHandler.guiId.MODULAR.ordinal());
+		super(TileOrbitalLaserDrill.class, GuiHandler.guiId.MODULAR.ordinal());
 		setTickRandomly(true).setUnlocalizedName("spaceLaser");
 	}
